refactor(binding-modbus): move modbus connectio to dedicated file
